requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==25.4.0
bcrypt==5.0.0
blinker==1.9.0
build==1.4.0
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
chromadb==1.5.5
click==8.1.8
coloredlogs==15.0.1
cryptography==46.0.5
dataclasses-json==0.6.7
datasets==4.5.0
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.17.0
durationpy==0.10
easyocr==1.7.2
eval_type_backport==0.3.1
exceptiongroup==1.3.1
filelock==3.19.1
FlashRank==0.2.10
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.46
googleapis-common-protos==1.73.0
grpcio==1.78.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==0.36.2
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
importlib_metadata==8.7.1
importlib_resources==6.5.2
instructor==1.14.5
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
langchain==0.3.28
langchain-chroma==0.2.6
langchain-community==0.3.31
langchain-core==0.3.83
langchain-experimental==0.3.4
langchain-huggingface==0.3.1
langchain-ollama==0.3.10
langchain-openai==0.3.35
langchain-text-splitters==0.3.11
langsmith==0.4.37
lazy-loader==0.5
markdown-it-py==3.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
mypy_extensions==1.1.0
narwhals==2.18.0
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.13.0
numpy==2.0.2
oauthlib==3.3.1
ollama==0.6.1
onnxruntime==1.19.2
openai==2.28.0
opencv-python-headless==4.13.0.92
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
orjson==3.11.5
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
plotly==6.6.0
propcache==0.4.1
protobuf==6.33.5
pyarrow==21.0.0
pybase64==1.4.3
pyclipper==1.3.0.post6
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.11.0
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pypdf==6.9.1
PyPika==0.51.1
pyproject_hooks==1.2.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
PyYAML==6.0.3
ragas==0.4.3
rank-bm25==0.2.2
referencing==0.36.2
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py==0.27.1
safetensors==0.7.0
scikit-image==0.24.0
scikit-learn==1.6.1
scikit-network==0.33.4
scipy==1.13.1
sentence-transformers==5.1.2
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
SQLAlchemy==2.0.48
streamlit==1.50.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2024.8.30
tiktoken==0.12.0
tokenizers==0.22.2
toml==0.10.2
tomli==2.4.0
torch==2.8.0
torchvision==0.23.0
tornado==6.5.5
tqdm==4.67.3
transformers==4.57.6
typer==0.23.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.39.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
